Fix small bug that in SET ANIMATION that prevented use of variables

+++ NPC Maker/NPC Maker/ScriptParser/ScriptHelpers.cs
@@ -522,7 +522,7 @@ namespace NPC_Maker.Scripts
             {
                 try
                 {
-                    outV.Value = (float)GetValueByType(SplitLine, Index, outV.Vartype, RangeMin, RangeMax);
+                    outV.Value = GetValueByType(SplitLine, Index, outV.Vartype, RangeMin, RangeMax);
                     return outV;
                 }
                 catch (Exception)
